services/gateway/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
psycopg[binary]>=3.1.8
starlette-exporter>=0.20.0
pydantic-settings>=2.4.0
structlog>=24.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate limiting
slowapi>=0.1.9

# Migrations
alembic>=1.13.2
PyYAML>=6.0.1
httpx>=0.27.0
nats-py>=2.7.2
temporalio>=1.7.0

# Tracing (optional)
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-exporter-otlp>=1.25.0

